Designed to download HARPS datasets..
Due to the fact that the HARPS archive times out after 2 hours and prevents further downloads, this program allows the user to download HARPS data in discrete segments for subsequent analysis. It also generates a CSV file containing all of the file paths, which can then be read by the subsequent analysis function.